payments/requirements.txt
flask
requests
sqlalchemy
kafka-python==2.0.2
prometheus_client